Timepix3Event2StdEventConverter: change stamp from unsigned long long int to unsigned int

+++ user/timepix3/module/src/Timepix3Event2StdEventConverter.cc
@@ -43,7 +43,7 @@ bool Timepix3TrigEvent2StdEventConverter::Converting(eudaq::EventSPC ev, eudaq::
   // 0110 1111 EEEE EEEE EEEE TTTT TTTT TTTT TTTT TTTT TTTT TTTT TTTT TTTF FFF0 0000
   // where: E=event_no, T=coarse_timestamp, F=fine_timestamp (0x0 to 0xc)
 
-  unsigned long long int stamp = (trigdata & 0x1E0) >> 5;
+  unsigned int stamp = (trigdata & 0x1E0) >> 5;
   long long int timestamp_raw = static_cast<long long int>(trigdata & 0xFFFFFFFFE00) >> 9;
   long long int timestamp = 0;
